incremented the "id" value for realm history and changed it from int to long

+++ app/src/main/java/edu/purdue/mehta109/trialanderror/EquationModel.java
@@ -12,7 +12,7 @@ public class EquationModel extends RealmObject{
         return numberID;
     }
 
-    public void setNumberID(int numberID) {
+    public void setNumberID(long numberID) {
         this.numberID = numberID;
     }
